Add files via upload
Same as before, hitting run example 1 will create the sample trip. It shows the route on the map and animates the car going along the route. The location of the car icon will need to be changed to wherever it is located on your system running it (Line 76). I also currently have the classes we added commented out if they have the timestamp as we haven't implemented that yet (I believe it's a json obect from one of the demos I was looking at).