Add files via upload
Should fix the packages, and startup for map to have initial position again. Example2 should actually  be working if it's not you need to download the gmaps.net.core v1.9.6 
We need the appointment list to be working. At minimum just pass up to 4 of the new appointments to the example function. It's been redone to take in lat/lng coordinates for the arguments and then performs normally with those.