Add files via upload
Run vans will run the inital 4 vans, was not able to get the loop or threading working to continue getting more vans going after the initial batch goes out. 
-Getting exceptions with the update position call within the loop. 0 exceptions with it commented out but exponential exceptions with it in.
**Commented out the hard coded longitude value in form 2 at line 250. test patients would not run without this line. I'm hoping the patients in the battle will run with it. 
made the firebase connection global hoping that would solve the exceptions on the update call but it only seems to lower it and not fix it.